refactor: extract broadcast choice code
They existed in RunForMoney3 plugin and extracted by accident (they should still remain in the codebase of that plugin) as I think these code existed only for that plugin.
But I'm not willing to remove these in this codebase as I'm lazy enough ha. Maybe they'll be moved back someday? I don't know.